+++ container-engine-lib/docs/changelog.md
@@ -9,6 +9,8 @@
   * All users of `container-engine-lib` need to upgrade to use this version
 * Refactor all functions in `kubernetes_kurtosis_backend_user_services_functions.go` to `user_service_functions` module
   * All users of `container-engine-lib` need to upgrade to use this version
+* Move `ServiceConfig` and `FilesArtifactsExpansion` out of `backend_interface` module into respect files in `objects` module
+  *   * All users of `container-engine-lib` need to upgrade to use this version
 
 ### Changes
 * Add `RegisterUserServices` and `StartUserServices` to `KurtosisBackend` interface